third_party/blink/web_tests/virtual/prefetch-sw/external/wpt/speculation-rules/prefetch/tentative/service-worker/basic.sub.https_origin=same-site_sw=fetch-handler-navigation-preload-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch with ServiceWorker (fetch-handler-navigation-preload)
  assert_equals: preloadResponse should be resolved expected (string) "resolved" but got (undefined) undefined
Harness: the test ran to completion.
